# src/services/FAQ/docs/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
  web.APIError:
    properties:
      createdAt:
        type: string
      errorCode:
        type: integer
      errorMessage:
        type: string
    type: object
  web.Pet:
    properties:
      category:
        properties:
          id:
            type: integer
          name:
            type: string
        type: object
      id:
        type: integer
      name:
        type: string
      photoUrls:
        items:
          type: string
        type: array
      status:
        type: string
      tags:
        items:
          type: Tag
        type: array
    type: object
  web.RevValue:
    properties:
      Data:
        type: integer
    type: object
host: '{{.Host}}'
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: '{{.Description}}'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: '{{.Title}}'
  version: '{{.Version}}'
paths:
  /file/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload file
      operationId: file.upload
      parameters:
      - description: this is a test file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: We need ID!!
          schema:
            $ref: '#/definitions/web.APIError'
            type: object
        "404":
          description: Can not find ID
          schema:
            $ref: '#/definitions/web.APIError'
            type: object
      summary: Upload file
  /testapi/get-string-by-int/{some_id}:
    get:
      consumes:
      - application/json
      description: get string by ID
      operationId: get-string-by-int
      parameters:
      - description: Some ID
        in: path
        name: some_id
        required: true
        type: integer
      - description: Some ID
        in: body
        name: some_id
        required: true
        schema:
          $ref: '#/definitions/web.Pet'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: We need ID!!
          schema:
            $ref: '#/definitions/web.APIError'
            type: object
        "404":
          description: Can not find ID
          schema:
            $ref: '#/definitions/web.APIError'
            type: object
      summary: Add a new pet to the store
  /testapi/get-struct-array-by-string/{some_id}:
    get:
      consumes:
      - application/json
      description: get struct array by ID
      operationId: get-struct-array-by-string
      parameters:
      - description: Some ID
        in: path
        name: some_id
        required: true
        type: string
      - description: Offset
        in: query
        name: offset
        required: true
        type: integer
      - description: Offset
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: We need ID!!
          schema:
            $ref: '#/definitions/web.APIError'
            type: object
        "404":
          description: Can not find ID
          schema:
            $ref: '#/definitions/web.APIError'
            type: object
swagger: "2.0"